+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Format title in public/index.html
 - Use title in Menu
 - Remove text input from Pause
+- renderTranscript() function
 
 ## 2.2.1 - May 25, 2025
 - Limit confirmation pseudo-element to terminal option buttons